first draft of job_verify
gonna leave this here for now, need to start summary paper
