+++ LIReC/config.py
@@ -10,7 +10,7 @@ configuration = {
     },
     'jobs_to_run': [
         ('poly_pslq', {
-            'args': { 'degree': (2, 1), 'bulk': 1000, 'filters': {
+            'args': { 'degree': 2, 'order': 1, 'bulk': 1000, 'filters': {
                 'global': { 'min_precision': 50 },
                 'PcfCanonical': { 'count': 2, 'balanced_only': True },
                 'Named': { 'count': 2 }
