+++ LIReC/jobs/job_poly_pslq.py
@@ -2,36 +2,38 @@
 Finds polynomial relations between constants in LIReC, using PSLQ.
 
 Configured as such:
-'degree': Tuple of the form (polydegree: int, innerdegree: int). All relations are structured like
-          multivariate polynomials over the constants and CFs, of degree polydegree with a maximum
-          exponent of innerdegree. For example, a 2-variable polynomial of degree (2,1) will be of
-          the form a+bx+cy+dxy (note the lack of x^2 and y^2), and a 4-variable polynomial of degree
-          (3,1) will be of the form:
-              a + bx+cy+dz+ew + fxy+gxz+hxw+iyz+jyw+kzw + lxyz+mxyw+nxzw+oyzw
-          Note here the lack of any single variable with an exponent greater than 1, and also the lack of xyzw.
-'bulk': How many of each "bulk type" to scan. A class of constants is
-        considered a bulk type iff we expect to have a lot of it in LIReC
-        (specified by whether or not it's in the BULK_TYPES array).
-'filters': A dictionary that specifies which kinds of constants to use to look for relations. Currently supported:
+'degree' + 'order':
+    Two integers. All relations are structured like multivariate polynomials over the constants and CFs,
+    of degree 'degree' with a maximum exponent of 'order'. For example, a 2-variable polynomial of
+    degree 2 and order 1 will be of the form a+bx+cy+dxy (note the lack of x^2 and y^2), and
+    a 4-variable polynomial of degree (3,1) will be of the form:
+        a + bx+cy+dz+ew + fxy+gxz+hxw+iyz+jyw+kzw + lxyz+mxyw+nxzw+oyzw
+    Note here the lack of any single variable with an exponent greater than 1, and also the lack of xyzw.
+'bulk':
+    How many of each "bulk type" to scan. A class of constants is
+    considered a bulk type iff we expect to have a lot of it in LIReC
+    (specified by whether or not it's in the BULK_TYPES array).
+'filters':
+    A dictionary that specifies which kinds of constants to use to look for relations. Currently supported:
     global filters: 'min_precision' specifies the minimal precision value of constants that will be used.
     'PcfCanonical': 'count' specifies how many pcf's at a time, 'balanced_only' filters to only PCFs of balanced degrees if set to True.
     'Named': 'count' specifies how many named constants at a time.
 '''
 import mpmath as mp
-from numpy import transpose
-from time import time
-from sqlalchemy import Integer, or_, Float
-from sqlalchemy.orm.attributes import flag_modified
-from sqlalchemy.sql.expression import func
+import os
+from functools import reduce
+from itertools import combinations, takewhile, product
 from logging import getLogger
 from logging.config import fileConfig
-from os import getpid
-from itertools import chain, combinations, combinations_with_replacement, takewhile, product
-from collections import Counter
-from functools import reduce
-from operator import mul
+from sqlalchemy import or_
+from sqlalchemy.sql.expression import func
+from sympy.polys.polytools import poly_from_expr
+from time import time
 from traceback import format_exc
+from LIReC.config import configuration
 from LIReC.lib import models, db_access
+from LIReC.lib.pcf import *
+from LIReC.lib.pslq_utils import *
 
 mp.mp.dps = 2000
 
@@ -45,7 +47,8 @@ BULK_SIZE = 500
 BULK_TYPES = {'PcfCanonical'}
 SUPPORTED_TYPES = ['Named', 'PcfCanonical']
 DEFAULT_CONST_COUNT = 1
-DEFAULT_DEGREE = (2, 1)
+DEFAULT_DEGREE = 2
+DEFAULT_ORDER = 1
 MIN_PRECISION_RATIO = 0.8
 MAX_PREC = 9999
 
@@ -67,27 +70,7 @@ def get_filters(filters, const_type):
 
     return filter_list 
 
-def poly_check(consts, exponents):
-    mp.mp.dps = min(c.base.precision for c in consts)
-    values = [mp.mpf(str(c.base.value)) for c in consts]
-    if 1 in values: # solely for backwards-compatibility. We don't need 1 in the DB!
-        return None, None, None
-    poly = [reduce(mul, (values[i] ** exp[i] for i in range(len(values))), mp.mpf(1)) for exp in exponents]
-    try:
-        mp.mp.dps = 15 # intentionally low-resolution to quickly try something basic...
-        res = mp.pslq(poly)
-        if res: # then calculating the substance in what we just found!
-            mp.mp.dps = max(c.base.precision for c in consts) + 10
-            min_prec = min(c.base.precision for c in consts)
-            return res, min(mp.floor(-mp.log10(abs(mp.fdot(poly, res)))), min_prec), min_prec
-    except ValueError:
-        # one of the constants has too small precision, or one constant
-        # is small enough that another constant is smaller than its precision.
-        # eitherway there's no relation to be found here!
-        pass 
-    return None, None, None
-
-def compress_relation(result, consts, exponents, degree):
+def compress_relation(result, consts, exponents, degree, order):
     # will need to use later, so evaluating into lists
     getLogger(LOGGER_NAME).debug(f'Original relation is {result}')
     
@@ -99,41 +82,35 @@ def compress_relation(result, consts, exponents, degree):
         redundant_coeffs |= set(indices_per_var[redundant_var])
         del consts[redundant_var]
     
-    polydegree, innerdegree = degree # remove redundant degrees
-    indices_per_polydegree = list(list(i for i, e in enumerate(exponents) if sum(e.values()) == j) for j in range(polydegree+1))
-    redundant_polydegrees = list(i for i, e in enumerate(indices_per_polydegree) if not any(result[j] for j in e))
-    redundant_polydegrees = list(takewhile(lambda x: sum(x) == polydegree, enumerate(sorted(redundant_polydegrees, reverse=True))))
-    if redundant_polydegrees:
-        polydegree = redundant_polydegrees[-1][1] - 1
-    redundant_coeffs.update(*indices_per_polydegree[polydegree+1:])
+    # remove redundant degrees and orders
+    indices_per_degree = list(list(i for i, e in enumerate(exponents) if sum(e.values()) == j) for j in range(degree+1))
+    redundant_degrees = list(i for i, e in enumerate(indices_per_degree) if not any(result[j] for j in e))
+    redundant_degrees = list(takewhile(lambda x: sum(x) == degree, enumerate(sorted(redundant_degrees, reverse=True))))
+    if redundant_degrees:
+        degree = redundant_degrees[-1][1] - 1
+    redundant_coeffs.update(*indices_per_degree[degree+1:])
     
-    indices_per_innerdegree = list(list(i for i, e in enumerate(exponents) if max(e.values(), default=0) == j) for j in range(innerdegree+1))
-    redundant_innerdegrees = list(i for i, e in enumerate(indices_per_innerdegree) if not any(result[j] for j in e))
-    redundant_innerdegrees = list(takewhile(lambda x: sum(x) == innerdegree, enumerate(sorted(redundant_innerdegrees, reverse=True))))
-    if redundant_innerdegrees:
-        innerdegree = redundant_innerdegrees[-1][1] - 1
-    redundant_coeffs.update(*indices_per_innerdegree[innerdegree+1:])
+    indices_per_order = list(list(i for i, e in enumerate(exponents) if max(e.values(), default=0) == j) for j in range(order+1))
+    redundant_orders = list(i for i, e in enumerate(indices_per_order) if not any(result[j] for j in e))
+    redundant_orders = list(takewhile(lambda x: sum(x) == order, enumerate(sorted(redundant_orders, reverse=True))))
+    if redundant_orders:
+        order = redundant_orders[-1][1] - 1
+    redundant_coeffs.update(*indices_per_order[order+1:])
     
-    degree = [polydegree, innerdegree]
-    getLogger(LOGGER_NAME).debug(f'True degree is {degree}')
+    getLogger(LOGGER_NAME).debug(f'True degree and order are {degree, order}')
     for i in sorted(redundant_coeffs, reverse=True):
         del result[i]
     
     getLogger(LOGGER_NAME).debug(f'Compressed relation is {result}')
 
-    return result, consts, degree
+    return result, consts, degree, order
 
-def get_exponents(degree, total_consts):
-    polydegree, innerdegree = degree
-    return list(c for c in map(Counter, chain.from_iterable(combinations_with_replacement(range(total_consts), i) for i in range(polydegree+1)))
-                if not any(i for i in c.values() if i > innerdegree))
-
-def relation_is_new(consts, degree, other_relations):
+def relation_is_new(consts, degree, order, other_relations):
     return not any(r for r in other_relations
-                   if {c.const_id for c in r.constants} <= {c.const_id for c in consts} and r.details[0] <= degree[0] and r.details[1] <= degree[1])
+                   if {c.const_id for c in r.constants} <= {c.const_id for c in consts} and r.details[0] <= degree and r.details[1] <= order)
 
-def check_consts(consts, exponents, degree):
-    result, true_prec, min_prec = poly_check(consts, exponents)
+def check_consts(consts, exponents, degree, order):
+    result, true_prec, min_prec = poly_check(consts, exponents = exponents)
     if not result:
         return []
     if DEBUG_PRINT_PRECISION_RATIOS:
@@ -144,21 +121,21 @@ def check_consts(consts, exponents, degree):
         if DEBUG_PRINT_PRECISION_RATIOS:
             print(f'Too low! Ignoring...')
         return []
-    result, new_consts, new_degree = compress_relation(result, consts, exponents, degree)
+    result, new_consts, new_degree, new_order = compress_relation(result, consts, exponents, degree, order)
     # now must check subrelations! PSLQ is only guaranteed to return a small norm,
     # but not guaranteed to return a 1-dimensional relation, see for example pslq([1,2,3])
     subrelations = []
     for i in range(1, len(consts)):
-        exponents = get_exponents(degree, i)
+        exponents = get_exponents(degree, order, i)
         for subset in combinations(consts, i):
-            subresult, true_prec2, min_prec2 = poly_check(subset, exponents)
+            subresult, true_prec2, min_prec2 = poly_check(subset, exponents = exponents)
             if subresult:
-                subresult, subconsts, subdegree = compress_relation(subresult, subset, exponents, degree)
-                if relation_is_new(subconsts, subdegree, subrelations) and true_prec2 / min_prec2 >= MIN_PRECISION_RATIO: # no need to check against new_relations + old_relations here btw
+                subresult, subconsts, subdegree, suborder = compress_relation(subresult, subset, exponents, degree, order)
+                if relation_is_new(subconsts, subdegree, suborder, subrelations) and true_prec2 / min_prec2 >= MIN_PRECISION_RATIO: # no need to check against new_relations + old_relations here btw
                     true_prec2 = min(true_prec2, MAX_PREC)
-                    subrelations += [models.Relation(relation_type=ALGORITHM_NAME, details=subdegree+subresult, precision=int(true_prec2), constants=[c.base for c in subconsts])]
+                    subrelations += [models.Relation(relation_type=ALGORITHM_NAME, details=[subdegree,suborder]+subresult, precision=int(true_prec2), constants=[c.base for c in subconsts])]
     true_prec = min(true_prec, MAX_PREC)
-    return subrelations if subrelations else [models.Relation(relation_type=ALGORITHM_NAME, details=new_degree+result, precision=int(true_prec), constants=[c.base for c in new_consts])]
+    return subrelations if subrelations else [models.Relation(relation_type=ALGORITHM_NAME, details=[new_degree,new_order]+result, precision=int(true_prec), constants=[c.base for c in new_consts])]
 
 def get_const_class(const_type):
     name = const_type + 'Constant'
@@ -171,8 +148,8 @@ def get_consts_from_query(const_type, query_data):
     return query_data[[i for i in range(len(query_data)) if isinstance(query_data[i][0], const_type)][0]]
 
 def get_consts(const_type, db, filters):
-    if const_type == 'Named':
-        return db.constants.join(models.Constant).filter(*get_filters(filters, const_type))
+    if const_type == 'Named': # Constant first intentionally! don't need extra details, but want to filter still
+        return db.session.query(models.Constant).join(models.NamedConstant).order_by(models.NamedConstant.const_id).filter(*get_filters(filters, const_type))    
 
 def run_query(filters=None, degree=None, bulk=None):
     fileConfig('LIReC/logging.config', defaults={'log_filename': 'pslq_const_manager'})
@@ -184,18 +161,18 @@ def run_query(filters=None, degree=None, bulk=None):
     bulk = bulk if bulk else BULK_SIZE
     getLogger(LOGGER_NAME).debug(f'Starting to check relations, using bulk size {bulk}')
     db = db_access.LIReC_DB()
-    results = [db.session.query(get_const_class(const_type)).join(models.Constant).filter(*get_filters(filters, const_type)).order_by(func.random()).limit(bulk).all() for const_type in bulk_types]
+    results = [db.session.query(models.Constant).join(get_const_class(const_type)).filter(*get_filters(filters, const_type)).order_by(func.random()).limit(bulk).all() for const_type in bulk_types]
     # apparently postgresql is really slow with the order_by(random) part,
     # but on 1000 CFs it only takes 1 second, which imo is worth it since
     # that allows us more variety in testing the CFs...
     # TODO what to do if results is unintentionally empty?
     db.session.close()
     getLogger(LOGGER_NAME).info(f'size of batch is {len(results) * bulk}')
-    return results # so pool_handler can correctly divide among the sub-processes
+    return results
 
-def execute_job(query_data, filters=None, degree=None, bulk=None, manual=False):
+def execute_job(query_data, filters=None, degree=None, order=None, bulk=None, manual=False):
     try: # whole thing must be wrapped so it gets logged
-        fileConfig('LIReC/logging.config', defaults={'log_filename': 'analyze_pcfs' if manual else f'pslq_const_worker_{getpid()}'})
+        fileConfig('LIReC/logging.config', defaults={'log_filename': 'analyze_pcfs' if manual else f'pslq_const_worker_{os.getpid()}'})
         global_filters = filters.get('global', {})
         filters.pop('global', 0) # instead of del so we can silently dispose of global even if it doesn't exist
         if not filters:
@@ -212,14 +189,15 @@ def execute_job(query_data, filters=None, degree=None, bulk=None, manual=False):
                 filters[const_type]['count'] = DEFAULT_CONST_COUNT
         total_consts = sum(c['count'] for c in filters.values())
         degree = degree if degree else DEFAULT_DEGREE
+        order = order if order else DEFAULT_ORDER
         getLogger(LOGGER_NAME).info(f'checking against {total_consts} constants at a time, subdivided into {({k : filters[k]["count"] for k in filters})}, using degree-{degree} relations')
-        if degree[0] > total_consts * degree[1]:
-            degree = (total_consts * degree[1], degree[1])
+        if degree > total_consts * order:
+            degree = total_consts * order
             getLogger(LOGGER_NAME).info(f'redundant degree detected! reducing to {degree}')
         
         db = db_access.LIReC_DB()
         subsets = [combinations(get_consts_from_query(const_type, query_data) if const_type in BULK_TYPES else get_consts(const_type, db, {**filters, 'global':global_filters}), filters[const_type]['count']) for const_type in filters]
-        exponents = get_exponents(degree, total_consts)
+        exponents = get_exponents(degree, order, total_consts)
         
         old_relations = db.session.query(models.Relation).all()
         orig_size = len(old_relations)
@@ -228,10 +206,10 @@ def execute_job(query_data, filters=None, degree=None, bulk=None, manual=False):
         # because finding new relations depends on the new relations we found so far!
         for consts in product(*subsets):
             consts = [c for t in consts for c in t] # need to flatten...
-            if relation_is_new(consts, degree, old_relations):
+            if relation_is_new(consts, degree, order, old_relations):
                 if DEBUG_PRINT_CONSTANTS:
                     getLogger(LOGGER_NAME).debug(f'checking consts: {[c.const_id for c in consts]}')
-                new_relations = check_consts(consts, exponents, degree)
+                new_relations = check_consts(consts, exponents, degree, order)
                 if new_relations:
                     getLogger(LOGGER_NAME).info(f'Found relation(s) on constants {[c.const_id for c in consts]}!')
                     try_count = 1
@@ -268,3 +246,159 @@ def summarize_results(results):
     if not all(results):
         getLogger(LOGGER_NAME).info(f'At least one of the workers had an exception! Check logs')
     getLogger(LOGGER_NAME).info(f'In total found {sum(r for r in results if r)} relations')
+
+
+PRECISION_LIMIT = 20
+
+def main(): # if testing PCFs that aren't already in LIReC, requires permissions to upload them!
+    os.makedirs(os.path.join(os.getcwd(), 'logs'), exist_ok=True)
+    config = [c for name, c in configuration if name == 'poly_pslq']
+    db = db_access.LIReC_DB()
+    pcfs = (PCF(*(poly_from_expr(poly)[0].all_coeffs() for poly in pcf)) for pcf in config['pcfs'])
+    successful, unsuccessful = db.add_pcfs(pcfs)
+    unsuccessful['Too imprecise'] = []
+    for key in unsuccessful:
+        if key == 'Already exist': # still want to test these...
+            canonical_forms = [[poly.all_coeffs() for poly in pcf.get_canonical_form()] for pcf in unsuccessful[key]]
+            for pcf in db.session.query(models.PcfCanonicalConstant):
+                if [pcf.P, pcf.Q] in canonical_forms: # TODO theoretically this can be implemented as a filter but i can't figure it out right now...
+                    if pcf.base.precision > PRECISION_LIMIT:
+                        successful += [pcf]
+                    else:
+                        unsuccessful['Too imprecise'] += [unsuccessful[key][canonical_forms.index([pcf.P, pcf.Q])]]
+        elif unsuccessful[key]:
+            polys = [[[int(c) for c in poly.all_coeffs()] for poly in pcf.get_canonical_form()] for pcf in unsuccessful[key]]
+            pcfs_string = reduce(lambda a,b: a+'\n\t'+str(b), polys[1:], str(polys[0]))
+            print(f'Could not add {len(unsuccessful[key])} pcfs with reason "{key}":\n\t{pcfs_string}\n')
+    db.session.close()
+    config = config['args']
+    if 'PcfCanonical' in config['filters'] and not successful:
+        print('None of the configued PCFs can be tested! Aborting...')
+    else: # bulk doesn't matter, it just exists due to the limitation of how EXECUTE_NEEDS_ARGS works
+        print(f'Testing relations with {len(successful)} PCFs')
+        execute_job(successful, config.get('filters', None), config.get('degree', None), config.get('order', None), manual = True)
+
+if __name__ == '__main__':
+    '''
+    Example config: (using BOINC zetas)
+    configuration = {
+        ...,
+        ('poly_pslq', {
+            'args': { ...
+            }, ...
+            'pcfs': [
+                ('2*n**5 + 7*n**4 + 14*n**3 + 16*n**2 + 9*n + 2', '-n**10 - 2*n**9 - n**8'),
+                ('2*n**5 + 7*n**4 + 19*n**3 + 25*n**2 + 16*n + 4', '-n**10 - 2*n**9 - n**8'),
+                ('2*n**5 + 7*n**4 + 19*n**3 + 30*n**2 + 21*n + 5', '-n**10 - 2*n**9 - n**8'),
+                ('2*n**5 + 8*n**4 + 12*n**3 + 10*n**2 + 5*n + 1', '-n**10 - 3*n**9 - 2*n**8'),
+                ('2*n**5 + 8*n**4 + 14*n**3 + 16*n**2 + 9*n + 2', '-n**10 - 3*n**9 - 2*n**8'),
+                ('2*n**5 + 8*n**4 + 18*n**3 + 22*n**2 + 13*n + 3', '-n**10 - 3*n**9 - 2*n**8'),
+                ('2*n**5 + 8*n**4 + 18*n**3 + 28*n**2 + 19*n + 6', '-n**10 - 3*n**9 - 2*n**8'),
+                ('2*n**5 + 8*n**4 + 24*n**3 + 34*n**2 + 23*n + 6', '-n**10 - 3*n**9 - 2*n**8'),
+                ('2*n**5 + 10*n**4 + 14*n**3 + 16*n**2 + 9*n + 2', '-n**10 - 5*n**9 - 4*n**8'),
+                ('2*n**5 + 10*n**4 + 26*n**3 + 34*n**2 + 21*n + 5', '-n**10 - 5*n**9 - 4*n**8'),
+                ('2*n**5 + 11*n**4 + 34*n**3 + 46*n**2 + 29*n + 7', '-n**10 - 6*n**9'),
+                ('2*n**5 + 13*n**4 + 10*n**3 + 10*n**2 + 5*n + 1', '-n**10 - 8*n**9'),
+                ('2*n**3 + 3*n**2 + 3*n + 1', '-n**6'),
+                ('2*n**3 + 3*n**2 + 11*n + 5', '-n**6'),
+                ('2*n**3 + 3*n**2 + 27*n + 13', '-n**6'),
+                ('2*n**3 + 3*n**2 + 51*n + 25', '-n**6'),
+                ('2*n**3 + 3*n**2 + 53*n + 26', '-n**6'),
+                ('2*n**3 + 3*n**2 + 81*n + 40', '-n**6'),
+                ('2*n**3 + 3*n**2 + 83*n + 41', '-n**6'),
+                ('2*n**3 + 3*n**2 + 123*n + 61', '-n**6'),
+                ('2*n**3 + 3*n**2 + 171*n + 85', '-n**6'),
+                ('2*n**3 + 3*n**2 + 227*n + 113', '-n**6'),
+                ('3*n**3 + 3*n**2 + 3*n + 1', '-2*n**6'),
+                ('3*n**3 + 6*n**2 + 6*n + 2', '-2*n**6'),
+                ('4*n**3 + 3*n**2 + 3*n + 1', '-3*n**6'),
+                ('4*n**3 + 9*n**2 + 9*n + 3', '-3*n**6'),
+                ('4*n**3 + 6*n**2 + 4*n + 1', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 6*n + 2', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 12*n + 5', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 22*n + 10', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 32*n + 15', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 36*n + 17', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 54*n + 26', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 76*n + 37', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 102*n + 50', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 106*n + 52', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 132*n + 65', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 162*n + 80', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 166*n + 82', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 204*n + 101', '-4*n**6'),
+                ('4*n**3 + 6*n**2 + 246*n + 122', '-4*n**6'),
+                ('5*n**3 + 3*n**2 + 3*n + 1', '-4*n**6'),
+                ('5*n**3 + 12*n**2 + 12*n + 4', '-4*n**6'),
+                ('5*n**3 + 9*n**2 + 9*n + 3', '-6*n**6'),
+                ('6*n**3 + 9*n**2 + 5*n + 1', '-n**6'),
+                ('6*n**3 + 3*n**2 + 3*n + 1', '-5*n**6'),
+                ('6*n**3 + 15*n**2 + 15*n + 5', '-5*n**6'),
+                ('6*n**3 + 6*n**2 + 6*n + 2', '-8*n**6'),
+                ('6*n**3 + 12*n**2 + 12*n + 4', '-8*n**6'),
+                ('6*n**3 + 9*n**2 + 9*n + 3', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 23*n + 10', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 33*n + 15', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 81*n + 39', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 115*n + 56', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 125*n + 61', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 153*n + 75', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 159*n + 78', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 233*n + 115', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 243*n + 120', '-9*n**6'),
+                ('6*n**3 + 9*n**2 + 249*n + 123', '-9*n**6'),
+                ('7*n**3 + 3*n**2 + 3*n + 1', '-6*n**6'),
+                ('7*n**3 + 18*n**2 + 18*n + 6', '-6*n**6'),
+                ('8*n**3 + 3*n**2 + 3*n + 1', '-7*n**6'),
+                ('8*n**3 + 21*n**2 + 21*n + 7', '-7*n**6'),
+                ('9*n**3 + 3*n**2 + 3*n + 1', '-8*n**6'),
+                ('9*n**3 + 24*n**2 + 24*n + 8', '-8*n**6'),
+                ('9*n**3 + 31*n**2 + 34*n + 12', '-8*n**6'),
+                ('10*n**3 + 3*n**2 + 3*n + 1', '-9*n**6'),
+                ('10*n**3 + 27*n**2 + 27*n + 9', '-9*n**6'),
+                ('34*n**3 + 51*n**2 + 27*n + 5', '-n**6'),
+                ('3*n**4 + 4*n**3 + 6*n**2 + 4*n + 1', '-2*n**8'),
+                ('3*n**4 + 8*n**3 + 12*n**2 + 8*n + 2', '-2*n**8'),
+                ('4*n**4 + 4*n**3 + 6*n**2 + 4*n + 1', '-3*n**8'),
+                ('4*n**4 + 12*n**3 + 18*n**2 + 12*n + 3', '-3*n**8'),
+                ('4*n**4 + 8*n**3 + 12*n**2 + 8*n + 2', '-4*n**8'),
+                ('4*n**4 + 8*n**3 + 20*n**2 + 16*n + 6', '-4*n**8'),
+                ('5*n**4 + 4*n**3 + 6*n**2 + 4*n + 1', '-4*n**8'),
+                ('5*n**4 + 16*n**3 + 24*n**2 + 16*n + 4', '-4*n**8'),
+                ('10*n**4 + 4*n**3 + 6*n**2 + 4*n + 1', '-9*n**8'),
+                ('10*n**4 + 36*n**3 + 54*n**2 + 36*n + 9', '-9*n**8'),
+                ('2*n**4 + 4*n**3 + 6*n**2 + 4*n + 1', '-n**8'),
+                ('2*n**4 + 4*n**3 + 10*n**2 + 8*n + 3', '-n**8'),
+                ('4*n**4 + 8*n**3 + 12*n**2 + 8*n + 2', '-4*n**8'),
+                ('4*n**4 + 8*n**3 + 20*n**2 + 16*n + 6', '-4*n**8'),
+                ('6*n**4 + 12*n**3 + 18*n**2 + 12*n + 3', '-9*n**8'),
+                ('6*n**4 + 12*n**3 + 30*n**2 + 24*n + 9', '-9*n**8'),
+                ('8*n**4 + 16*n**3 + 24*n**2 + 16*n + 4', '-16*n**8'),
+                ('8*n**4 + 16*n**3 + 40*n**2 + 32*n + 12', '-16*n**8'),
+                ('16*n**5 + 40*n**4 + 50*n**3 + 35*n**2 + 13*n + 2', '-64*n**10'),
+                ('16*n**5 + 40*n**4 + 56*n**3 + 44*n**2 + 18*n + 3', '-64*n**10'),
+                ('2*n**5 + 5*n**4 + 10*n**3 + 10*n**2 + 5*n + 1', '-n**10'),
+                ('2*n**5 + 5*n**4 + 22*n**3 + 28*n**2 + 15*n + 3', '-n**10'),
+                ('2*n**5 + 5*n**4 + 42*n**3 + 58*n**2 + 45*n + 13', '-n**10'),
+                ('16*n**5 + 40*n**4 + 120*n**3 + 140*n**2 + 90*n + 23', '-64*n**10'),
+                ('-16*n**5 - 40*n**4 - 120*n**3 - 140*n**2 - 90*n - 23', '-64*n**10'),
+                ('-2*n**5 - 5*n**4 - 42*n**3 - 58*n**2 - 45*n - 13', '-n**10'),
+                ('16*n**5 + 40*n**4 + 50*n**3 + 35*n**2 + 13*n + 2', '-64*n**10'),
+                ('16*n**5 + 40*n**4 + 56*n**3 + 44*n**2 + 18*n + 3', '-64*n**10'),
+                ('-2*n**5 - 5*n**4 - 22*n**3 - 28*n**2 - 23*n - 7', '-n**10'),
+                ('-2*n**5 - 5*n**4 - 10*n**3 - 10*n**2 - 5*n - 1', '-n**10'),
+                ('2*n**5 + 5*n**4 + 22*n**3 + 28*n**2 + 15*n + 3', '-n**10'),
+                ('-16*n**5 - 40*n**4 - 56*n**3 - 44*n**2 - 18*n - 3', '-64*n**10'),
+                ('-16*n**5 - 40*n**4 - 50*n**3 - 35*n**2 - 13*n - 2', '-64*n**10'),
+                ('2*n**5 + 5*n**4 + 42*n**3 + 58*n**2 + 45*n + 13', '-n**10'),
+                ('16*n**5 + 40*n**4 + 120*n**3 + 140*n**2 + 90*n + 23', '-64*n**10'),
+                ('2*n**5 + 5*n**4 + 10*n**3 + 10*n**2 + 5*n + 1', '-n**10'),
+                ('2*n**5 + 5*n**4 + 22*n**3 + 28*n**2 + 15*n + 3', '-n**10'),
+                ('2*n**5 + 5*n**4 + 22*n**3 + 28*n**2 + 23*n + 7', '-n**10'),
+                ('2*n**5 + 5*n**4 + 42*n**3 + 58*n**2 + 45*n + 13', '-n**10')
+            ]
+        })
+        ...
+    }
+    '''
+    main()
